tests/wpt/tests/largest-contentful-paint/web-font-styled-text-resize-swap-smaller.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @font-face {
    font-family: 'TestFaceSmaller';
    src: url('/fonts/CSSTest/csstest-basic-regular.ttf?pipe=trickle(d0.5)');
    size-adjust: 30%;
    font-display: swap;
  }

  .test {
    font-family: 'TestFaceSmaller';
  }

</style>
<!--
  Web-font styled text that gets resized during swap period should not make a
  LCP emission if the new size is smaller than the current LCP element size.
-->
<div class="test">LCP: Web Font Styled Text Resize</div>
<script>
  function LCPEntryList(t) {
    return new Promise(resolve => {
      let = lcpEntries = [];
      new PerformanceObserver((entryList, observer) => {
        lcpEntries = lcpEntries.concat(entryList.getEntries());
        if (lcpEntries) {
          // Adding timeout to wait a bit more so that if there are more than
          // expected LCP entries emitted, they can be observed.
          t.step_timeout(() => {
            resolve(lcpEntries);
            observer.disconnect();
          }, 200);
        }
      }).observe({ type: 'largest-contentful-paint', buffered: true });
    });
  }

  promise_test(async t => {
    await document.fonts.ready;

    // Verify web font is loaded.
    assert_own_property(window, 'PerformanceResourceTiming', "ResourceTiming not supported");
    let url = '/fonts/CSSTest/csstest-basic-regular.ttf?pipe=trickle(d0.5)';
    var absoluteURL = new URL(url, location.href).href;
    assert_equals(performance.getEntriesByName(absoluteURL).length, 1, 'Web font should be downloaded');

    // Verify web font is available.
    assert_true(document.fonts.check('10px TestFaceSmaller'), 'Font should be the web font added');

    // Verify there is only one LCP entry.
    let entryList = await LCPEntryList(t);
    assert_equals(entryList.length, 1, 'Web font styled text resize that occurs during swap period but is smaller should not make a new LCP emission.')

  }, "LCP should be not updated if the web font styled text resizes to be smaller during the swap period");
</script>
